requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
google-generativeai==0.8.5
python-dotenv
fpdf==1.7.2
flask-bcrypt
psycopg2-binary
flask-sqlalchemy
Flask-SQLAlchemy==3.1.1

